refactor tensor library by splitting lib.rs into modules
