+++ tensor/src/tensor/ops/mod.rs
@@ -1,0 +1,4 @@
+pub mod binary;
+pub mod movement;
+pub mod reduce;
+pub mod unary;
